feat(utils): Single implementation to fetch debug ids
