+++ packages/core/src/utils/prepareEvent.ts
@@ -6,10 +6,16 @@ import type {
   EventHint,
   Scope as ScopeInterface,
   ScopeContext,
-  StackFrame,
   StackParser,
 } from '@sentry/types';
-import { GLOBAL_OBJ, addExceptionMechanism, dateTimestampInSeconds, normalize, truncate, uuid4 } from '@sentry/utils';
+import {
+  addExceptionMechanism,
+  dateTimestampInSeconds,
+  getFilenameToDebugIdMap,
+  normalize,
+  truncate,
+  uuid4,
+} from '@sentry/utils';
 
 import { DEFAULT_ENVIRONMENT } from '../constants';
 import { getGlobalScope } from '../currentScopes';
@@ -161,51 +167,12 @@ function applyClientOptions(event: Event, options: ClientOptions): void {
   }
 }
 
-const debugIdStackParserCache = new WeakMap<StackParser, Map<string, StackFrame[]>>();
-
 /**
  * Puts debug IDs into the stack frames of an error event.
  */
 export function applyDebugIds(event: Event, stackParser: StackParser): void {
-  const debugIdMap = GLOBAL_OBJ._sentryDebugIds;
-
-  if (!debugIdMap) {
-    return;
-  }
-
-  let debugIdStackFramesCache: Map<string, StackFrame[]>;
-  const cachedDebugIdStackFrameCache = debugIdStackParserCache.get(stackParser);
-  if (cachedDebugIdStackFrameCache) {
-    debugIdStackFramesCache = cachedDebugIdStackFrameCache;
-  } else {
-    debugIdStackFramesCache = new Map<string, StackFrame[]>();
-    debugIdStackParserCache.set(stackParser, debugIdStackFramesCache);
-  }
-
   // Build a map of filename -> debug_id
-  const filenameDebugIdMap = Object.entries(debugIdMap).reduce<Record<string, string>>(
-    (acc, [debugIdStackTrace, debugIdValue]) => {
-      let parsedStack: StackFrame[];
-      const cachedParsedStack = debugIdStackFramesCache.get(debugIdStackTrace);
-      if (cachedParsedStack) {
-        parsedStack = cachedParsedStack;
-      } else {
-        parsedStack = stackParser(debugIdStackTrace);
-        debugIdStackFramesCache.set(debugIdStackTrace, parsedStack);
-      }
-
-      for (let i = parsedStack.length - 1; i >= 0; i--) {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        const stackFrame = parsedStack[i]!;
-        if (stackFrame.filename) {
-          acc[stackFrame.filename] = debugIdValue;
-          break;
-        }
-      }
-      return acc;
-    },
-    {},
-  );
+  const filenameDebugIdMap = getFilenameToDebugIdMap(stackParser);
 
   try {
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
