+++ packages/utils/src/index.ts
@@ -40,3 +40,4 @@ export * from './buildPolyfills';
 export * from './propagationContext';
 export * from './vercelWaitUntil';
 export * from './version';
+export * from './debug-ids';
